% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosePorts.R
\name{choosePorts}
\alias{choosePorts}
\title{Give the user the option of choose the working ports writing them in an
emergent window}
\usage{
choosePorts()
}
\value{
the code of the selected working ports
}
\description{
Give the user the option of choose the working ports writing them in an
emergent window
}
