% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_catches_xlsx.R
\name{create_catches_xlsx}
\alias{create_catches_xlsx}
\title{Create xlsx file with data of catches as a resume to check the information.}
\usage{
create_catches_xlsx(catches, catches_in_lengths, year, month, path = getwd())
}
\arguments{
\item{year}{year of the data. This is used only to name the exported files.
This function doesn't filter by year.}

\item{month}{month of the data. This is used only to name the exported files.
This function doesn't filter by month.}

\item{path}{path where the 'catches' file is located and where the exported
files will be saved.}

\item{catches:}{catches data frame returned by the importRIMCatches() function.}

\item{catches_in_lengths:}{catches in lengths data frame returned by the
importRIMCatchesInLengths() function.}
}
\description{
Create xlsx file with data of catches as a resume to check the information.
}
