% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_check_files_xlsx.R
\name{create_check_files_xlsx}
\alias{create_check_files_xlsx}
\title{Function to create a xls file with the checked samples from SIRENO's database}
\usage{
create_check_files_xlsx(capturas_tot, tallas, cod_ports)
}
\arguments{
\item{cod_ports}{vector with the code ports to filter.}

\item{capturas_tot:}{data frame returned by the importRIMCatches() or
importRIMFiles() functions.}

\item{tallas:}{lengths data frame returned by the importRIMCatchesInLengths () or
importRIMFiles() functions.}
}
\value{
export all the previous xls files generated
}
\description{
Function to create a xls file with the checked samples from SIRENO's database
}
