% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_headers_xlsx.R
\name{create_headers_xlsx}
\alias{create_headers_xlsx}
\title{Create header file to facilitate the manual sampling check.}
\usage{
create_headers_xlsx(lengths, year, month, path = getwd())
}
\arguments{
\item{lengths}{Lengths data frame from Sireno report.}

\item{path}{Path where lengths file is located and to save the file.}
}
\value{
Export xlsx file with headers resume.
}
\description{
Create header file to facilitate the manual sampling check.
}
