% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lengths_xls.R
\name{create_lengths_xls}
\alias{create_lengths_xls}
\title{Function to create a xls file with the lengths sampling document from SIRENO's database}
\usage{
create_lengths_xls(lens)
}
\arguments{
\item{lens:}{lengths data frame returned by the importRIMLengths() or
importRIMFiles() functions.}
}
\value{
a new table in xls format to see the data in a diferent format
where de values of the columns turns to the rows and vice versa
}
\description{
Function to create a xls file with the lengths sampling document from SIRENO's database
}
