% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_catches_xls.R
\name{createCatchesXls}
\alias{createCatchesXls}
\title{Function to create a xls file with the catches sampling document from SIRENO's database}
\usage{
createCatchesXls(lens, catches)
}
\arguments{
\item{catches:}{data frame returned by the importRIMCatches() or
importRIMFiles() functions.}

\item{lens:}{lengths data frame returned by the importRIMLengths() or
importRIMFiles() functions.}
}
\value{
a new table in xls format to see the data in a diferent format
where de values of the columns turns to the rows and vice versa
}
\description{
Function to create a xls file with the catches sampling document from SIRENO's database
}
