% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCheckFilesXlsx.R
\name{createCheckFilesXlsx}
\alias{createCheckFilesXlsx}
\title{Function to create a xls file with the checked samples from SIRENO's database}
\usage{
createCheckFilesXlsx(capturas_tot, tallas)
}
\arguments{
\item{capturas_tot:}{data frame returned by the importRIMCatches() or
importRIMFiles() functions.}

\item{tallas:}{lengths data frame returned by the importRIMCatchesInLengths () or
importRIMFiles() functions.}
}
\value{
export all the previous xls files generated
}
\description{
Function to create a xls file with the checked samples from SIRENO's database
}
