% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_ports.R
\name{encode_ports}
\alias{encode_ports}
\title{This function gives you the codes of the ports what you are working.
You have two ways: using a vector with the name of the ports or insert
them by an emergent window}
\usage{
encode_ports(work_ports, dialog_box = FALSE)
}
\arguments{
\item{work_ports}{vector with the names of our working ports}

\item{dialog_box}{boolean paramater to choose if you work by the vector or the
emergent window way. By default is defined with FALSE value, because the main way of
work is through a vector. Then, if you want to work choosing the ports you only
need to declare dialog_box with TRUE}
}
\value{
the code of the selected working ports
}
\description{
This function gives you the codes of the ports what you are working.
You have two ways: using a vector with the name of the ports or insert
them by an emergent window
}
