% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeadersXls.R
\name{createHeadersXls}
\alias{createHeadersXls}
\title{Create header file to facilitate the manual sampling check.}
\usage{
createHeadersXls(lens)
}
\arguments{
\item{lens}{Lengths data frame of Sireno report.}
}
\value{
Export xlsx file with headers resume.
}
\description{
Create header file to facilitate the manual sampling check.
}
