% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_ports.R
\name{encode_ports}
\alias{encode_ports}
\title{Return the codes of the given ports.
There are two ways: using a vector with the name of the ports or select
them by an emergent window.}
\usage{
encode_ports(ports, dialog_box = FALSE)
}
\arguments{
\item{ports}{vector with the names of ports}

\item{dialog_box}{boolean parameter. If TRUE, the emergent window is showed.
If FALSE, the 'ports' argument must be given. FALSE by default.}
}
\value{
vector with the code of the ports.
}
\description{
Return the codes of the given ports.
There are two ways: using a vector with the name of the ports or select
them by an emergent window.
}
